app/celery_worker.py
import os
import requests
import google.generativeai as genai
from celery import Celery
from dotenv import load_dotenv

# Load environment variables from .env file
load_dotenv()

# Configure Celery
celery_app = Celery("worker", broker=os.getenv("REDIS_URL"))

# Configure Gemini
genai.configure(api_key=os.getenv("GEMINI_API_KEY"))

@celery_app.task(bind=True, max_retries=3, default_retry_delay=30)
def generate_audio_task(self, text: str, phone_number: str):
    """
    This is the background task that does the actual work.
    """
    try:
        print(f"Processing task for: {phone_number}")

        # This is a placeholder for the actual Gemini Text-to-Speech API call.
        # You will need to replace this with the correct method from the google-generativeai library
        # based on their documentation for TTS.
        # Let's assume the result is binary audio data.
        # Example: audio_data = genai.text_to_speech(...)

        # --- Placeholder for audio data ---
        print("Simulating Gemini API call...")
        audio_data = f"This is a fake audio file for {phone_number} with text: {text}".encode('utf-8')
        file_name = f"{phone_number}.mp3"
        # --- End of placeholder ---

        # Send the audio file to the success webhook
        success_url = os.getenv("SUCCESS_WEBHOOK_URL")
        files = {'audio_file': (file_name, audio_data, 'audio/mpeg')}

        print(f"Sending file to success webhook: {success_url}")
        response = requests.post(success_url, files=files)
        response.raise_for_status() # Will raise an error for 4xx/5xx responses

        return {"status": "success", "phone_number": phone_number}

    except Exception as exc:
        print(f"Task failed for {phone_number}. Error: {exc}. Retrying...")
        # Retry the task with a countdown. If max_retries is reached, it will raise an error.
        # Celery handles the error propagation.
        raise self.retry(exc=exc)