requirements.txt
fastapi
uvicorn[standard]
celery
redis
google-generativeai
requests
python-dotenv